Refactorizando crear reglas, fin iteración #1
