+++ memoria/src/routes/createrules/+page.svelte
@@ -1,14 +1,17 @@
 <script lang="ts">
+    // Importamos las librerias necesarias
     import { onMount } from "svelte";
     import { goto } from "$app/navigation";
-    import { XMLParser } from "fast-xml-parser";
-    import { writable, get, type Writable } from "svelte/store";
+    import { writable, get } from "svelte/store";
     import "../types";
     import { themeStore } from "../../stores";
     import {
         fileUploadBpmn,
         fileUploadContext,
     } from "../../functions/importdata";
+    import * as functionRulecreation from "../../functions/rulecreation";
+
+    // Variables
     let selectedAction = "";
     let selectedAction1 = "Delete Action";
     let selectedAction2 = "";
@@ -23,6 +26,21 @@
     let showModalEliminar = false;
     let rule_selected: Rule, subrule_selected: Rule;
     let activities: activity[] = [];
+    let divElement: HTMLElement;
+    let yOffset = 0;
+
+    let xmlContext: string = "";
+    let xmlBpmn: string = "";
+    let attributesContext: any[] = [];
+
+    let activity: activity = {
+        id: 0,
+        name: "Default Activity - No activity selected",
+        subname: "",
+        rules: [],
+    };
+
+    let rules = writable<Rule[]>([]);
 
     function toggleModalNivel1() {
         showModal = !showModal;
@@ -45,134 +63,27 @@
         showModalEliminar = !showModalEliminar;
     }
 
-    let rules = writable<Rule[]>([]);
-
-    function getUniqueId(): string {
-        const characters =
-            "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-        let result = "";
-        const charactersLength = characters.length;
-        for (let i = 0; i < 20; i++) {
-            result += characters.charAt(
-                Math.floor(Math.random() * charactersLength),
-            );
-        }
-        return result;
-    }
-
-    function addConnectorRule(parentComplexRuleId?: string) {
-        const newRule: ConnectorRule = {
-            id: getUniqueId(),
-            type: "Conector",
-            logical_operator: "Or",
-            logicals: ["And", "Or"],
-        };
-        updateRuleInComplex(newRule, parentComplexRuleId);
-    }
-
     function addSimpleRule(parentComplexRuleId?: string, id?: string) {
-        if (parentComplexRuleId != null) {
-            if (shouldAddConnector(parentComplexRuleId)) {
-                addConnectorRule(parentComplexRuleId);
-            }
-        } else {
-            if (shouldAddConnectorAtTopLevel()) {
-                addConnectorRule();
-            }
-        }
-        const newRule: SimpleRule = {
-            id: getUniqueId(),
-            type: "Simple",
-            numberRule: get(rules).length + 1,
-            attribute: "",
-            value: "",
-            attributes: attributesContext.map((attr) => attr.Attribute),
-            values: [],
-        };
-        updateRuleInComplex(newRule, parentComplexRuleId);
+        const result = functionRulecreation.addSimpleRule(
+            rules,
+            attributesContext,
+            parentComplexRuleId,
+            id,
+        );
+        rules = result;
     }
 
     function addComplexRule(parentComplexRuleId?: string) {
-        if (parentComplexRuleId != null) {
-            if (shouldAddConnector(parentComplexRuleId)) {
-                addConnectorRule(parentComplexRuleId);
-            }
-        } else {
-            if (shouldAddConnectorAtTopLevel()) {
-                addConnectorRule();
-            }
-        }
-        const newRule: ComplexRule = {
-            id: getUniqueId(),
-            type: "Complex",
-            numberRule: get(rules).length + 1,
-            rules: [],
-        };
-        updateRuleInComplex(newRule, parentComplexRuleId);
+        const result = functionRulecreation.addComplexRule(
+            rules,
+            parentComplexRuleId,
+        );
+        rules = result;
     }
 
     function deleteRuleById(ruleId: string) {
-        rules.update((currentRules) => {
-            return removeRule(currentRules, ruleId);
-        });
-    }
-
-    function removeRule(rulesList: Rule[], ruleId: string): Rule[] {
-        for (let i = 0; i < rulesList.length; i++) {
-            const rule = rulesList[i];
-            if (rule.id === ruleId) {
-                // Check if previous rule is a connector and remove it
-                if (i > 0 && rulesList[i - 1].type === "Conector") {
-                    rulesList.splice(i - 1, 2);
-                } else if (
-                    i < rulesList.length - 1 &&
-                    rulesList[i + 1].type === "Conector"
-                ) {
-                    rulesList.splice(i, 2);
-                } else {
-                    rulesList.splice(i, 1);
-                }
-                return rulesList;
-            } else if (rule.type === "Complex") {
-                rule.rules = removeRule(rule.rules, ruleId);
-            }
-        }
-        return rulesList;
-    }
-
-    function shouldAddConnector(parentComplexRuleId: string): boolean {
-        const parentRule = findRuleById(get(rules), parentComplexRuleId);
-        return (
-            parentRule !== null &&
-            parentRule.type === "Complex" &&
-            parentRule.rules.length > 0 &&
-            parentRule.rules.slice(-1)[0].type !== "Conector"
-        );
-    }
-
-    function shouldAddConnectorAtTopLevel(): boolean {
-        const topLevelRules = get(rules);
-        return (
-            topLevelRules.length > 0 &&
-            topLevelRules.slice(-1)[0].type !== "Conector"
-        );
-    }
-
-    function updateRuleInComplex(
-        newRule: Rule,
-        parentComplexRuleId: string | undefined,
-    ) {
-        if (parentComplexRuleId != null) {
-            rules.update((rs) => {
-                let parentRule = findRuleById(rs, parentComplexRuleId);
-                if (parentRule && parentRule.type === "Complex") {
-                    parentRule.rules.push(newRule);
-                }
-                return rs;
-            });
-        } else {
-            rules.update((currentRules) => [...currentRules, newRule]);
-        }
+        const result = functionRulecreation.deleteRuleById(rules, ruleId);
+        rules = result;
     }
 
     function findRuleById(rules: Rule[], id: string): Rule | null {
@@ -189,20 +100,8 @@
         }
         return null;
     }
-    let divElement: HTMLElement;
-    let yOffset = 0;
 
-    let xmlContext: string = "";
-    let xmlBpmn: string = "";
-    let attributesContext: any[] = [];
-
-    let activity: activity = {
-        id: 0,
-        name: "Default Activity - No activity selected",
-        subname: "",
-        rules: [],
-    };
-    // Efecto parallax, ademas, recolectaremos los datos contenidos en el localstorage
+    // Recolectaremos los datos contenidos en el localstorage
     onMount(() => {
         var activitySelect = localStorage.getItem("activitySelect")!;
         if (activitySelect != null) {
@@ -225,6 +124,7 @@
                 }
             }
         }
+        // Efecto parallax
         const handleScroll = () => {
             const scrollPosition = divElement.scrollTop; // Cambiamos a usar scrollTop del div
             yOffset = scrollPosition * 0.5; // Ajusta el valor para controlar la velocidad del efecto parallax
@@ -241,16 +141,6 @@
             divElement.removeEventListener("scroll", handleScroll);
         };
     });
-
-    // Creando una instancia del parser y del builder
-    const parserOptions = {
-        ignoreAttributes: false,
-        attributeNamePrefix: "",
-        allowBooleanAttributes: true,
-        parseNodeValue: true,
-        parseAttributeValue: true,
-    };
-    const parser = new XMLParser(parserOptions);
 
     // Función para manejar la carga de archivos contexto organizacional
     async function handleFileUploadContext() {
@@ -280,6 +170,7 @@
         replaceaction = activities.map((task) => task.name);
     }
 
+    // Función para guardar los cambios del atributo en las reglas
     function attributeRuleSelected(event: Event, rule_selected: SimpleRule) {
         let attributeSelect = (event.target as HTMLSelectElement).value;
         // Buscamos la regla seleccionada y modificamos el valor del atributo
@@ -301,6 +192,7 @@
         });
     }
 
+    // Función para guardar los cambios del valor en las reglas
     function valueRuleSelected(event: Event, rule_selected: SimpleRule) {
         let valueSelect = (event.target as HTMLSelectElement).value;
         // Buscamos la regla seleccionada y modificamos el valor del atributo
@@ -315,6 +207,7 @@
         });
     }
 
+    // Función para guardar los cambios del conector en las reglas
     function conectorRuleSelected(event: Event, rule_selected: ConnectorRule) {
         let conectorSelect = (event.target as HTMLSelectElement).value;
         // Buscamos la regla seleccionada y modificamos el valor del atributo
@@ -327,6 +220,7 @@
         });
     }
 
+    // Función para guardar los cambios de todas las reglas en el localStorage
     function addRulesLocalStorage() {
         // Obtenemos las actividades que estan en el localStorage
         var task = localStorage.getItem("taskNames")!;
@@ -362,23 +256,11 @@
         // Guardamos el objeto con las reglas en el localStorage
         localStorage.setItem("taskNames", JSON.stringify(jsonTask));
     }
-    function clearAllRules(): void {
-        var tasks = JSON.parse(localStorage.getItem("taskNames")!);
-        var activitySelect = JSON.parse(
-            localStorage.getItem("activitySelect")!,
-        );
-        tasks.forEach((task: any) => {
-            if (task.id === activitySelect.id) {
-                {
-                    task.rules = [];
-                    task.replaceActivity = undefined;
-                    task.deleted = undefined;
-                    task.replaced = undefined;
-                }
-            }
-        });
-        localStorage.setItem("taskNames", JSON.stringify(tasks));
-        activities = tasks;
+
+    // Limpiar todas las reglas realizadas
+    function clearAllRules() {
+        const result = functionRulecreation.clearAllRules(activities);
+        activities = result;
     }
 </script>
